Rename module and cleanup code
