+++ medical_prescription_disease_state/views/medical_patient_disease_view.xml
@@ -19,7 +19,7 @@
                         name="%(medical_prescription.medical_prescription_order_line_action)d"
                         type="action" class="oe_stat_button" icon="fa-list"
                         context="{'search_default_disease_id': active_id}">
-                        <field name="count_prescription_order_line_ids"
+                        <field name="count_prescription_order_lines"
                             widget="statinfo" string="Prescription Order Lines"/>
                     </button>
                 </xpath>
@@ -38,11 +38,11 @@
                 ref="medical_patient_disease.medical_patient_view_disease_tree"/>
             <field name="arch" type="xml">
                 <xpath expr="/tree" position="attributes">
-                    <attribute name="colors">blue:last_prescription_order_line_state == 'inactive'</attribute>
+                    <attribute name="colors">blue:last_prescription_order_line_active == False</attribute>
                 </xpath>
                 <xpath expr="//field[@name='state']" position="after">
                     <field name="last_prescription_order_line_id" />
-                    <field name="last_prescription_order_line_state" invisible="1" />
+                    <field name="last_prescription_order_line_active" invisible="1" />
                 </xpath>
             </field>
         </record>
